test(context): split tests.rs into scenario submodules
src/context/tests.rs was 970 lines, above U-16's 800-line hard ceiling.
Split it along scenario lines:

- tests/render.rs (264 lines): render_recent_sessions / render_memory_index
  / render_workstreams / render_core_memory tests
- tests/load.rs (394 lines): load_context_data + ContextLimits env tests
- tests/sessions.rs (188 lines): query_recent_summaries tests
- tests/mod.rs (130 lines): shared fixture helpers

Closes #77
